refactor(frontend): components moved and variables renamed
