+++ packages/frontend/components/Onboarding/ExploreCarousel.tsx
